Refactor step functions
